(1) Change in Preview panel on the Structure tab of the Designer: 	(a) Refresh icon exchanged for refresh tool button; 	(b) Header of the Preview panel is bigger and it aligns with surrounding Y-axis panel's headers bottom edge; 	(c) Preview panel's title ("Preview" is aligned left on the panel's header
(2) Plus tool button on the Y-axis panel is gathered with the gear tool button (in its header)

(3) For chart type combo box items (when combo is opened) bold font style is removed. Font of item names (name of chart types) now is normal (no style). Also, chart type name is now aligned with the icon when combo is opened

(4) Possibility for the right part of the Designer (the one that contains tab panel with its toolbar) to resize according to current browser's (windows) dimensions is retrieved