(1) FIXED KNOWAGE-64 ISSUE
(2) "Line offset" parameters removed from the Additional axis parameters field set inside of the Axis style popup for the GAUGE chart. This parameter is managed by the velocity model.

(3) "Delete all" option provided also for the Plotbands fields set in Axis style popup in the form of delete icon tool. This option provides opportunity for user to remove all plots that he set for the GAUGE chart document.

(4) Phase 1 of reconfiguration of the Step 2 (Configuration) tab panel of the Designer. All panels are split by their current (old) appearance and spread through all the width of the Designer's right part (width of the tab panel).

(5) Introduction of the ChartConfigurationRoot.js file (class) that all panels on the Step 2 (Configuration) will extend. This root panel will initially contain only the margin property for all panels, since it is their common feature.