Connected to the KNOWAGE-269 REQUIREMENT:
If there is too much data in the Highcharts chart, this library does not let us render such a chart. This is because the "turboThreshold" property (a property inside plotOptions.series property inside of the Velocity Model of that chart) is set to its default value of 1000. This means that chart will not render if the number of data is more than 1000. 

Current (temporary) solution: for Highcharts chart types BAR, LINE, RADAR, PIE, SCATTER, TREEMAP and HEATMAP we increased the value for 'turboThreshold' parameter to 5000 since this will fit more to real-life scenarios.

Requirement: Improve current implementation of rendering the chart when we have more data than possible (showing the info text for empty message (no data)), so to have a e.g. preceding window popup by which we will inform the user to specify dataset (that current chart should use) with smaller number of records (less data) in order to render chart properly.