(1) Defined missing names of some Velocity Model files that are connected to their chart types.
(2) In Plotbands field set inside of the Axis style configuration popup for the GAUGE chart color field is improved. Instead of displaying the hexadecimal value of the color picked for particular plot, the color itself is displayed.

(3) When adding new Y-axis panel for the GAUGE chart, current chart style is applied to it.