Color palette improved:
(1) Reordering of colors (items) enabled
(2) The height of the panel that contains the palette (grid) is dynamically changed, depending on the action user takes (removing of item(s) or adding a single item)
(3) Name of the column of the grid (palette) is changed to "Color" (from "Font color"). This bug is fixed.