(1) "Text alignment" combobox in the Legend panel Designer's Configuration tab is hidden, since it does not make any change when chosen.
(2) The "Opposite" checkbox in the Axis style configuration popup (for the Y-axis of the SCATTER chart type) is disabled since it made a lot of complications when user combine other parameters that affect the axis label position in a way that those labels are not visible anymore or at least are moved out of the chart's visible part.

(3) The "Back" button for drilling up on the HEATMAP chart is positioned at the bottom center part of the chart, in order to fix its position. This way we prevent overlapping of the chart's title+subtitle or chart itself with the button.