FIXED, INTRODUCED OR IMPROVED:
(1) Introduce of two modalities for displaying (formatting) of the particular serie item. Now in the "Serie style configuration" popup for certain serie item there is a combo that offers to user to choose between those two modalities: Format or Precision. According to the one that is chosen appropriate GUI elements will be displayed/hidden. The "Format" string should satisfy the expression form that Highcharts library prescribes and in it's base it is equal to the already existing and provided combination of precision value, thousands separator sign (this one is hidden in the VM), prefix and postfix strings.

(2) For chart types BAR, LINE and RADAR the data labels and tooltip formatter are improved so they can handle these new features. For BAR and LINE chart the tooltip and data labels  improvements are specific since there we can experience drilling down, so we need to forward the tooltip configuration of the main serie (item that we drilled) towards its subcomponents. This is needed since the old implementation did not take the formatting of the serie to lower levels.

(3) This improvement, but slightly different than the one for the BAR, LINE and RADAR, is done also for the GAUGE and SCATTER charts (in their VM). In addition, the Format field that was initialy defined for the GAUGE chart in its "Data labels" fieldset is removed, so now it takes the format value from the serie customization ("Format or precision" combo and its belonging fields in the "Serie" fieldset).

(4) For purpose of using the Highchart Export web application as the engine that provides us the preview and export features for Highcharts chart we added a new input parameter when calling the VM from the Designer: 'exportWebApp' of type boolean. This one is provided within the Designer.js, on the place where we handle clicking on the "refresh" button within the header of the Preview panel on the "Structure" tab of the Designer page. This information will be useful for the VM of the chart that can handle drilldown and drillup (BAR, LINE), since for those there is specific part of the VM code that does not go well with the Export web application, since it does not need it. We need to skip these lines of code in order to provide functionality of previewing and exporting of the Highcharts chart.

(5) Improving validation for series and categories for some chart types.

(6) The "Template HTML" text area is removed from the Tooltip fieldset of the Serie style configuration popup since it is not used anywhere.

(7) The left container (panel) of the Designer's main page (the one that contains comboboxes for chart's type and style, as well as measure and attribute grid containers) is now resizable.

(8) The empty text for axis panels on the Designer's Structure tab is improved and it is specific for each chart type.

(9) For the left Y-axis panel (the main one for the Y-axes) and for the bottom X-axis panel there is a new tool that is provided - information button in the their headers (the one with a question mark icon). This tool when clicked provide the custom empty text that is specific for the panel for current chart type (the text that appears in the panels body when the panel is empty - not a single item is dropped in it).

(10) The maximum number for series for the PIE chart is 4 (four). User is prevented from dropping the fifth serie item. Up to this number of series the chart can be rendered.

(11) When user clicks on the "remove all items in the panel" icon in the header of the appropriate serie/category panel, the popup appears with the question if the user is sure that he wants to remove all items that are in the panel.

(12) The problem with the bottom X-axis panel's height not fitting the actual number of already existing document is fixed, so now also for this case (opening already existing document whose number number of categories exceeded the number of items for which do not need to provide an additional space for dropping new item(s)).