(1) The chart type combo box in the Designer is now communicating (synchronized) with the service (with the URL appending "1.0/pages/types") that provides him the necessary information that it gets from the "engine_config.xml" - all chart types that are available in the Designer: 	[a] The service URL appending is changed: from the original "types" to "1.0/pages/types" inside the ChartTypeResource.java. 	[b] Images of chart types are now renamed (they take the same name as the name of the chart type, all in lower case - e.g. for BAR chart type, the name of the picture is "bar.png") 		and put into new location in the project/server ("WebContent/img/designer/chart/types" in the project). This is needed for the chart type combo box. 	[c] Store of the combo box is now populated dynamically, since we are waiting for the answer of the REST call that we make when the Designer page is loading. When the data is 		received we are taking them and instantly generating the store of the combo box with all necessary fields.
(2) Tooltip is provided for series and categories that are placed in their axes containers. It is displayed whenever a mouse is over one item.

(3) Fixed the problem of missing data about series' absolute value or percentage format ('serieShowAbsValue' and 'serieShowPercentage' properties) from the data (chart) model. Change is done inside of the ChartColumnsContainerManager.js file.

(4) The maximum font size for the WORDCLOUD chart type is increased from 200px to 500px, since sometimes the most valuable word is many times bigger by its value than others, so those are almost invisible in that case. By increasing this value, they can be more visible.