[Ticket#2021021987000027] Streaming window size is now configurable
The dimension of the window impacts on the memory used by the server when performing XLSX export. Remember that crosstable widgets can be exported as excel pivot tables only if the number of rows is smaller than the streaming window size, otherwise only raw data will be exported (as they come from the data service, without totals or formatting).