[KNOWAGE-7390][KNOWAGE-7395]
KNOWAGE-7390
Added a method that will read the width, and apply it to the table cells.
PrimeVue Table has an issue where column re-size is not working well with pre-set widths while the table is scrollable. Scroll is the main issue here....

KNOWAGE-7395
Fixed scrolling and added a stable scroll on datable iself, so layout should be consistent.
As mentioned above...scrolling is really messing up the table styles.
fd9576