[KNOWAGE-8196] Refactoring OAuth2 filter
The OAuth2 filter is now always on but its actual management of the
request is defined by settings via system properties or environment
variable. By default, the OAuth2 don't do anything but with the right
set of configs it changes its behavior.

The OAuth2 config is now managed in the servlet filter init method: if
something is wrong the entire application won't boot up.